+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.12.1] - 2022-07-22
+
+### Added
+
+- Added a warning message when the customer schema is invalid.
+
 ## [1.12.0] - 2022-07-21
 
 ### Added
@@ -24,7 +30,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed on update the user data, and it disappeared from the list
 
 ## [1.11.2] - 2022-07-06
+
 ### Fixed
+
 - Fixed organization bulk action button text.
 
 ## [1.11.1] - 2022-07-04
@@ -34,6 +42,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial Crowdin integration
 
 ## [1.11.0] - 2022-06-23
+- 
+### Added
+- UI support for organization `tradeName` and cost center `phoneNumber` fields
 
 ## [1.10.3] - 2022-06-22
 
@@ -43,7 +54,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- UI support for organization `tradeName` and cost center `phoneNumber` fields
 - Github Action to trigger manual tests by dispatch
 
 ## [1.10.2] - 2022-06-14